Домашнее задание для занятия 27
Создан класс EndpointConfig.java с сервисом Уточка
В пакет clients вынесен класс DuckActionsTest c запросами для каждого эндпоинта
Переписаны тесты с использованием методов из DuckActionsTest
