+++ src/test/java/autotests/tests/duck_action_controller/DuckActionFlyTest.java
@@ -1,6 +1,6 @@
 package autotests.tests.duck_action_controller;
 
-import autotests.DuckActionsTest;
+import autotests.clients.DuckActionsTest;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
@@ -9,7 +9,6 @@ import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
 
 import static com.consol.citrus.container.FinallySequence.Builder.doFinally;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
 public class DuckActionFlyTest extends DuckActionsTest {
     @Test(description = "Проверить полет уточки с активными крыльями")
@@ -55,12 +54,5 @@ public class DuckActionFlyTest extends DuckActionsTest {
 
         String responseMessage = generateMessageJson("Wings are not detected");
         validateResponse(runner, HttpStatus.INTERNAL_SERVER_ERROR, responseMessage);
-    }
-
-    public void duckFly(TestCaseRunner runner, String id) {
-        runner.$(http().client(getServerUrl())
-                .send()
-                .get("/api/duck/action/fly")
-                .queryParam("id", id));
     }
 }
