+++ src/test/java/autotests/tests/duck_action_controller/DuckActionPropertiesTest.java
@@ -1,6 +1,6 @@
 package autotests.tests.duck_action_controller;
 
-import autotests.DuckActionsTest;
+import autotests.clients.DuckActionsTest;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
@@ -9,7 +9,6 @@ import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
 
 import static com.consol.citrus.container.FinallySequence.Builder.doFinally;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
 public class DuckActionPropertiesTest extends DuckActionsTest {
     @Test(description = "Проверить получение свойств деревянной уточки с целым четным ID")
@@ -52,12 +51,5 @@ public class DuckActionPropertiesTest extends DuckActionsTest {
 
         String responseMessage = generateDuckJson(color, height, material, sound, wingsState);
         validateResponse(runner, HttpStatus.OK, responseMessage);
-    }
-
-    public void duckProperties(TestCaseRunner runner, String id) {
-        runner.$(http().client(getServerUrl())
-                .send()
-                .get("/api/duck/action/properties")
-                .queryParam("id", id));
     }
 }
