+++ src/test/java/autotests/tests/duck_action_controller/DuckActionSwimTest.java
@@ -1,6 +1,6 @@
 package autotests.tests.duck_action_controller;
 
-import autotests.DuckActionsTest;
+import autotests.clients.DuckActionsTest;
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
@@ -9,7 +9,6 @@ import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
 
 import static com.consol.citrus.container.FinallySequence.Builder.doFinally;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
 public class DuckActionSwimTest  extends DuckActionsTest {
     @Test(description = "Проверить плавание существующей уточки")
@@ -43,12 +42,5 @@ public class DuckActionSwimTest  extends DuckActionsTest {
                         "  \"message\": \"Duck with id = " + duckId() + " is not found\",\n" +
                         "  \"path\": \"/api/duck/action/fly\"\n" +
                         "}");
-    }
-
-    public void duckSwim(TestCaseRunner runner, String id) {
-        runner.$(http().client(getServerUrl())
-                .send()
-                .get("/api/duck/action/swim")
-                .queryParam("id", id));
     }
 }
